Move handling of non-existing resources
Non-existing resources will be handled in the service layer.
If a resource is requested via API call, the service
layer is responsible to throw an exception if this can be found.
The router layer is only responsible to translate this into an
HTTP status code. This is done by a global exception handler.